backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0

# Data Processing & ETL
pandas>=2.2.0
numpy>=1.24.0
pyarrow>=14.0.0

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis>=5.0.0
alembic>=1.12.0

# Configuration & Environment
python-dotenv==1.0.0
pydantic-settings>=2.0.0

# Monitoring & Observability
prometheus-client>=0.18.0
structlog>=23.2.0

# Data Quality & Validation
great-expectations>=0.17.0
cerberus>=1.3.4

# Schema Management
jsonschema>=4.19.0

# Async & Streaming
aiofiles>=23.2.1
asyncio-mqtt>=0.16.1

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.1
faker>=19.12.0